refactor: reminder의 mandalart에 대한 의존 수정
