+++ src/main/java/com/koa/koalamailman/mandalart/application/MandalartUseCase.java
@@ -5,6 +5,7 @@ import com.koa.koalamailman.mandalart.application.dto.MandalartDto;
 import com.koa.koalamailman.mandalart.infrastructure.GoalRepository;
 import com.koa.koalamailman.mandalart.infrastructure.MandalartRepository;
 import com.koa.koalamailman.mandalart.domain.Goal;
+import com.koa.koalamailman.mandalart.domain.GoalLevel;
 import com.koa.koalamailman.mandalart.domain.Mandalart;
 import com.koa.koalamailman.global.exception.error.MandalartErrorCode;
 import com.koa.koalamailman.global.exception.BusinessException;
@@ -12,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.*;
 
 @Service
@@ -67,5 +69,22 @@ public class MandalartUseCase {
 
         if (!Objects.equals(mandalart.getUserId(), userId)) throw new BusinessException(MandalartErrorCode.MANDALART_FORBIDDEN);
         return mandalart;
+    }
+
+    @Transactional(readOnly = true)
+    public List<Mandalart> findDueMandalarts(LocalDateTime until) {
+        return mandalartRepository.findDueReminders(until);
+    }
+
+    @Transactional(readOnly = true)
+    public List<Goal> getCoreAndMainGoals(Long mandalartId) {
+        return goalRepository.findByMandalartIdAndLevelIn(mandalartId, List.of(GoalLevel.CORE, GoalLevel.MAIN));
+    }
+
+    @Transactional
+    public void rescheduleReminder(Long mandalartId, LocalDateTime nextTime) {
+        Mandalart mandalart = mandalartRepository.findById(mandalartId).orElse(null);
+        if (mandalart == null) return;
+        mandalart.getReminderOption().setRemindScheduledAt(nextTime);
     }
 }
