+++ src/main/java/com/koa/koalamailman/reminder/presentation/dto/request/ReminderOptionRequest.java
@@ -1,6 +1,6 @@
 package com.koa.koalamailman.reminder.presentation.dto.request;
 
-import com.koa.koalamailman.reminder.domain.RemindInterval;
+import com.koa.koalamailman.mandalart.domain.RemindInterval;
 
 import java.time.LocalDateTime;
